requirements.txt
Django==2.2.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
whitenoise==4.1.2
